clean and rename files
